feat(redis): add batch size option
