+++ db/redhat.go
@@ -9,6 +9,7 @@ import (
 	"github.com/inconshreveable/log15"
 	"github.com/knqyf263/gost/models"
 	"github.com/knqyf263/gost/util"
+	"github.com/spf13/viper"
 	pb "gopkg.in/cheggaaa/pb.v1"
 	"gorm.io/gorm"
 )
@@ -167,7 +168,7 @@ func (r *RDBDriver) deleteAndInsertRedhat(conn *gorm.DB, cves []models.RedhatCVE
 		return fmt.Errorf("Failed to delete old records. err: %s", errs.Error())
 	}
 
-	for idx := range chunkSlice(len(cves), r.batchSize) {
+	for idx := range chunkSlice(len(cves), viper.GetInt("batch-size")) {
 		if err = tx.Create(cves[idx.From:idx.To]).Error; err != nil {
 			return fmt.Errorf("Failed to insert. err: %s", err)
 		}
